Now enables statsd to either act as normally or as a slave.
Slave are effectively reverse proxies which will wait for a tcp connection from a master and forward all traffic immediately to that master.  Reconnections are handled automatically.

This enables slaves sat outside your normal firewall to talk to a master inside your firewall which can then collate as normal and push to graphite.  This has been done to enable remote collectors without having to open up inbound connections through the firewall as the master initiates the connection